cmd: init
This also adds a generate command for docs and completion.
